# practica1-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi Tienda Arch API",
        "description": "API REST para gestión de tienda Vue + Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión de usuario",
                "operationId": "da2ca4d7a5cdec77b8b45a20373535cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "nanche@nanche.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "nanche"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login correcto, devuelve el token"
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    }
                }
            }
        },
        "/api/v1/categorias": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Obtener lista de categorías",
                "operationId": "97e1a0f988ccfca6a410696722979d54",
                "responses": {
                    "200": {
                        "description": "Devuelve la lista de categorías disponibles"
                    }
                }
            }
        },
        "/api/v1/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener lista paginada de productos",
                "operationId": "2af0fdbcd7c972b8f9bd45123cd57cfb",
                "responses": {
                    "200": {
                        "description": "Devuelve la lista de productos disponibles"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "434c8599d4fe9aaa5aee59a5be0cf3fd",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Forzar respuesta JSON",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio",
                                    "stock",
                                    "categoria_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Teclado Mecánico Keychron"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Teclado inalámbrico"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500.5
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 18
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación en los datos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/productos/{producto}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener los detalles de un producto específico",
                "operationId": "06e92bcd80e312c48c1e1ce4e330b215",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID numérico del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del producto obtenidos exitosamente"
                    },
                    "404": {
                        "description": "Producto no encontrado en la base de datos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto existente",
                "operationId": "862d152c7fdbe73364442d3c652a015f",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto a modificar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Teclado Mecánico V2"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Versión actualizada con switches blue"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1800
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 45
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 18
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado correctamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación en los datos enviados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto del catálogo",
                "operationId": "b781643d3d24541ec0e8f1e8fa487ec7",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado exitosamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum-token",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Categorías",
            "description": "Categorías"
        },
        {
            "name": "Productos",
            "description": "Productos"
        }
    ]
}